various pedantic fixes various memory fixes fixed a few tests

+++ src/core/CLucene/util/VoidList.h
@@ -1,7 +1,7 @@
 /*------------------------------------------------------------------------------
 * Copyright (C) 2003-2006 Ben van Klinken and the CLucene Team
-* 
-* Distributable under the terms of either the Apache License (Version 2.0) or 
+*
+* Distributable under the terms of either the Apache License (Version 2.0) or
 * the GNU Lesser General Public License, as specified in the COPYING file.
 ------------------------------------------------------------------------------*/
 #ifndef _lucene_util_VoidList_
@@ -16,7 +16,7 @@ CL_NS_DEF(util)
 * A template to encapsulate various list type classes
 * @internal
 */
-template<typename _kt,typename _base,typename _valueDeletor> 
+template<typename _kt,typename _base,typename _valueDeletor>
 class CLUCENE_INLINE_EXPORT __CLList:public _base,LUCENE_BASE {
 private:
 	bool dv;
@@ -48,8 +48,8 @@ public:
 		into[i] = NULL;
 	}
 
-	void set(int32_t i, _kt val) {
-    if ( dv && i < base::size() ) 
+	void set(size_t i, _kt val) {
+    if ( dv && i < base::size() )
 			_valueDeletor::doDelete((*this)[i]);
     if ( i+1 > base::size() ) base::resize(i+1);
 		(*this)[i] = val;
@@ -59,11 +59,11 @@ public:
 	void delete_back(){
 		if ( base::size() > 0 ){
 			iterator itr = base::end();
-			if ( itr != base::begin()) 
+			if ( itr != base::begin())
 				itr --;
 			_kt key = *itr;
 			base::erase(itr);
-			if ( dv ) 
+			if ( dv )
 				_valueDeletor::doDelete(key);
 		}
 	}
@@ -73,7 +73,7 @@ public:
 			iterator itr = base::begin();
 			_kt key = *itr;
 			base::erase(itr);
-			if ( dv ) 
+			if ( dv )
 				_valueDeletor::doDelete(key);
 		}
 	}
@@ -89,20 +89,20 @@ public:
 		base::clear();
 	}
 
-	void remove(int32_t i, bool dontDelete=false){
+	void remove(size_t i, bool dontDelete=false){
 	    if ( i < base::size() ){
 		  iterator itr=base::begin();
 		  itr+=i;
 		  _kt key = *itr;
 		  base::erase( itr );
-		  if ( dv && !dontDelete ) 
+		  if ( dv && !dontDelete )
 			  _valueDeletor::doDelete(key);
 	    }
 	}
 	void remove(iterator itr, bool dontDelete=false){
 		_kt key = *itr;
 		base::erase( itr );
-		if ( dv && !dontDelete ) 
+		if ( dv && !dontDelete )
 			_valueDeletor::doDelete(key);
 	}
 };
@@ -112,7 +112,7 @@ public:
 //growable arrays of Objects (like a collection or list)
 //a list, so can contain duplicates
 //it grows in chunks... todo: check jlucene for initial size of array, and growfactors
-template<typename _kt, typename _valueDeletor=CL_NS(util)::Deletor::Dummy> 
+template<typename _kt, typename _valueDeletor=CL_NS(util)::Deletor::Dummy>
 class CLUCENE_INLINE_EXPORT CLVector:public __CLList<_kt, CL_NS_STD(vector)<_kt> , _valueDeletor>
 {
 public:
@@ -126,21 +126,21 @@ public:
 //a list, so can contain duplicates
 //*** a very simple list - use <valarray>
 //(This class is roughly equivalent to Vector, except that it is unsynchronized.)
-#define CLArrayList CLVector 
+#define CLArrayList CLVector
 #define CLHashSet CLHashList
 #define CLList CLVector
 
-//implementation of the List interface, provides access to the first and last list elements in O(1) 
+//implementation of the List interface, provides access to the first and last list elements in O(1)
 //no comparator is required... and so can contain duplicates
 //a simple list with no comparator
 //*** a very simple list - use <list>
 #ifdef LUCENE_DISABLE_HASHING
-   #define CLHashList CLSetList 
+   #define CLHashList CLSetList
 #else
 
 template<typename _kt,
 	typename _Comparator=CL_NS(util)::Compare::TChar,
-	typename _valueDeletor=CL_NS(util)::Deletor::Dummy> 
+	typename _valueDeletor=CL_NS(util)::Deletor::Dummy>
 class CLUCENE_INLINE_EXPORT CLHashList:public __CLList<_kt, CL_NS_HASHING(_CL_HASH_SET)<_kt,_Comparator> , _valueDeletor>
 {
 public:
@@ -151,7 +151,7 @@ public:
 };
 #endif
 
-template<typename _kt, typename _valueDeletor=CL_NS(util)::Deletor::Dummy> 
+template<typename _kt, typename _valueDeletor=CL_NS(util)::Deletor::Dummy>
 class CLUCENE_INLINE_EXPORT CLLinkedList:public __CLList<_kt, CL_NS_STD(list)<_kt> , _valueDeletor>
 {
 public:
@@ -162,7 +162,7 @@ public:
 };
 template<typename _kt,
 	typename _Comparator=CL_NS(util)::Compare::TChar,
-	typename _valueDeletor=CL_NS(util)::Deletor::Dummy> 
+	typename _valueDeletor=CL_NS(util)::Deletor::Dummy>
 class CLUCENE_INLINE_EXPORT CLSetList:public __CLList<_kt, CL_NS_STD(set)<_kt,_Comparator> , _valueDeletor>
 {
 public:
